Modularized 3 tasks, created global state struct
