+++ components/stratum/CMakeLists.txt
@@ -1,3 +1,12 @@
-idf_component_register(SRCS "utils.c" "mining.c" "stratum_api.c"
-                    INCLUDE_DIRS "include"
-                    REQUIRES json mbedtls)
+idf_component_register(
+SRCS 
+    "utils.c" 
+    "mining.c" 
+    "stratum_api.c"
+                    
+INCLUDE_DIRS 
+    "include"
+REQUIRES 
+    json
+    mbedtls
+)
