+++ main/CMakeLists.txt
@@ -1,12 +1,19 @@
-idf_component_register(SRCS
-"adc.c"
-"DS4432U.c"
-"EMC2101.c"
-"fonts.c"
-"INA260.c"
-"led_controller.c"
-"miner.c"
-"oled.c"
-"system.c"
-"work_queue.c"
-INCLUDE_DIRS ".")
+idf_component_register(
+SRCS
+    "adc.c"
+    "DS4432U.c"
+    "EMC2101.c"
+    "fonts.c"
+    "INA260.c"
+    "led_controller.c"
+    "miner.c"
+    "oled.c"
+    "system.c"
+    "work_queue.c"
+    "./tasks/stratum_task.c"
+    "./tasks/create_jobs_task.c"
+    "./tasks/asic_task.c"
+INCLUDE_DIRS 
+    "."
+    "tasks"
+)
