+++ main/tasks/asic_task.h
@@ -1,0 +1,6 @@
+#ifndef ASIC_TASK_H_
+#define ASIC_TASK_H_
+
+void ASIC_task(void * pvParameters);
+
+#endif
