+++ main/tasks/create_jobs_task.h
@@ -1,0 +1,6 @@
+#ifndef CREATE_JOBS_TASK_H_
+#define CREATE_JOBS_TASK_H_
+
+void create_jobs_task(void * pvParameters);
+
+#endif
