+++ main/tasks/stratum_task.h
@@ -1,0 +1,6 @@
+#ifndef STRATUM_TASK_H_
+#define STRATUM_TASK_H_
+
+void stratum_task(void * pvParameters);
+
+#endif
